feat: introduce tokens directory
