+++ packages/design-system/components/button/Button.styled.js
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled"
 import { css } from "@emotion/react"
-import theme from "design-system/theme"
+import theme from "../../tokens/theme"
 
 const sizes = {
   small: {
