+++ packages/design-system/components/illustration/Illustration.jsx
@@ -1,6 +1,6 @@
 import { oneOf } from "prop-types"
 
-import HeartWithDogFace from "design-system/illustrations/heartWithDogFace.svg"
+import HeartWithDogFace from "design-system/tokens/illustrations/heartWithDogFace.svg"
 
 const sizes = {
   heartWithDogFace: {
