+++ packages/design-system/patterns/textBanner/TextBanner.jsx
@@ -1,7 +1,7 @@
 import { string, oneOf, node } from "prop-types"
 
 import Typography from "design-system/components/typography"
-import theme from "design-system/theme"
+import theme from "../../tokens/theme"
 
 import * as Styled from "./TextBanner.styled"
