+++ packages/design-system/test-utils.jsx
@@ -2,7 +2,7 @@ import { render } from "@testing-library/react"
 import { ThemeProvider } from "@emotion/react"
 import { node } from "prop-types"
 
-import theme from "design-system/theme"
+import theme from "./tokens/theme"
 
 const Providers = ({ children }) => (
   <ThemeProvider theme={theme}>{children}</ThemeProvider>
