--- packages/design-system/theme/index.js
@@ -1,3 +1,0 @@
-import theme from "./theme"
-
-export default theme
