+++ packages/design-system/tokens/colors/__stories__/Colors.stories.mdx
@@ -1,0 +1,24 @@
+import { Meta, ColorPalette, ColorItem } from "@storybook/addon-docs"
+
+import colors from "design-system/tokens/colors"
+
+<Meta
+  title="Tokens/Colors"
+  parameters={{
+    previewTabs: {
+      canvas: { hidden: true },
+    },
+  }}
+/>
+
+# Colors
+
+A global colors palette available for all the compontens via theme.
+
+<ColorPalette>
+  {Object.entries(colors).map((palette) => (
+    <ColorItem key={palette[0]} title={palette[0]} colors={palette[1]} />
+  ))}
+</ColorPalette>
+
+---
