+++ packages/design-system/tokens/colors/index.js
@@ -1,0 +1,41 @@
+const colors = {
+  primary: {
+    100: "#FCDDDF",
+    200: "#F9BBBF",
+    300: "#F5989F",
+    400: "#EF5460",
+    500: "#C1121F",
+    600: "#A0151F",
+  },
+  neutrals: {
+    100: "#FFFEFE",
+    200: "#FEF9F8",
+    300: "#FBEDE8",
+    400: "#F8DFD8",
+    500: "#F4CEC3",
+  },
+  grey: {
+    100: "#EBEBEB",
+    200: "#DFDFDF",
+    300: "#C4C4C4",
+    400: "#818181",
+    500: "#414141",
+    600: "#0C0C0C",
+    700: "#020202",
+  },
+  complementary: {
+    100: "#FAF4F4",
+    200: "#F0DBDA",
+  },
+  blue: {
+    100: "#0D99FF",
+  },
+  success: {
+    100: "#14AE5C",
+  },
+  error: {
+    100: "#DB3541",
+  },
+}
+
+export default colors
