+++ packages/design-system/tokens/icons/__stories__/Icons.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, IconGallery, IconItem } from "@storybook/addon-docs"
 
 import { icons } from "design-system/components/icon"
 
-<Meta title="Icons" />
+<Meta title="Tokens/Icons" />
 
 # Icons gallery
