+++ packages/design-system/tokens/illustrations/__stories__/Illustrations.stories.mdx
@@ -2,7 +2,7 @@ import { Meta, IconGallery, IconItem } from "@storybook/addon-docs"
 
 import { illustrations } from "design-system/components/illustration"
 
-<Meta title="Illustrations" />
+<Meta title="Tokens/Illustrations" />
 
 # Illustrations gallery
