+++ packages/design-system/tokens/theme/index.js
@@ -1,0 +1,12 @@
+import typography, { fontFamily } from "../typography"
+import colors from "../colors"
+import breakpoints from "../breakpoints"
+
+const theme = {
+  fontFamily,
+  typography,
+  colors,
+  breakpoints,
+}
+
+export default theme
