+++ packages/design-system/tokens/typography/index.js
@@ -1,0 +1,173 @@
+export const fontFamily = "'Poppins', sans-serif"
+
+const typography = {
+  desktop: {
+    h1: {
+      fontWeight: 600,
+      fontSize: "60px",
+      lineHeight: "120%",
+      letterSpacing: "-1px",
+    },
+    h2: {
+      fontWeight: 600,
+      fontSize: "48px",
+      lineHeight: "120%",
+      letterSpacing: "-1px",
+    },
+    h3: {
+      fontWeight: 600,
+      fontSize: "28px",
+      lineHeight: "120%",
+      letterSpacing: "-0.5px",
+    },
+    h4: {
+      fontWeight: 500,
+      fontSize: "24px",
+      lineHeight: "125%",
+    },
+    h5: {
+      fontWeight: 500,
+      fontSize: "20px",
+      lineHeight: "120%",
+    },
+    h6: {
+      fontWeight: 500,
+      fontSize: "16px",
+      lineHeight: "130%",
+    },
+    bodyLarge: {
+      fontWeight: 300,
+      fontSize: "28px",
+      lineHeight: "140%",
+    },
+    bodyTitle: {
+      fontWeight: 300,
+      fontSize: "20px",
+      lineHeight: "140%",
+      letterSpacing: "-0.5px",
+    },
+    bodySmall: {
+      fontWeight: 300,
+      fontSize: "16px",
+      lineHeight: "140%",
+    },
+    bodyTiny: {
+      fontWeight: 300,
+      fontSize: "14px",
+      lineHeight: "140%",
+    },
+  },
+  tablet: {
+    h1: {
+      fontWeight: 600,
+      fontSize: "48px",
+      lineHeight: "125%",
+      letterSpacing: "-1px",
+    },
+    h2: {
+      fontWeight: 600,
+      fontSize: "40px",
+      lineHeight: "120%",
+      letterSpacing: "-1px",
+    },
+    h3: {
+      fontWeight: 600,
+      fontSize: "30px",
+      lineHeight: "150%",
+      letterSpacing: "-0.5px",
+    },
+    h4: {
+      fontWeight: 500,
+      fontSize: "28px",
+      lineHeight: "125%",
+      letterSpacing: "-0.5px",
+    },
+    h5: {
+      fontWeight: 500,
+      fontSize: "22px",
+      lineHeight: "150%",
+    },
+    h6: {
+      fontWeight: 500,
+      fontSize: "18px",
+      lineHeight: "150%",
+    },
+    bodyLarge: {
+      fontWeight: 300,
+      fontSize: "22px",
+      lineHeight: "145%",
+    },
+    bodyTitle: {
+      fontWeight: 300,
+      fontSize: "20px",
+      lineHeight: "145%",
+    },
+    bodySmall: {
+      fontWeight: 300,
+      fontSize: "16px",
+      lineHeight: "150%",
+    },
+    bodyTiny: {
+      fontWeight: 300,
+      fontSize: "14px",
+      lineHeight: "140%",
+    },
+  },
+  mobile: {
+    h1: {
+      fontWeight: 600,
+      fontSize: "40px",
+      lineHeight: "100%",
+      letterSpacing: "-1px",
+    },
+    h2: {
+      fontWeight: 600,
+      fontSize: "36px",
+      lineHeight: "110%",
+      letterSpacing: "-0.5px",
+    },
+    h3: {
+      fontWeight: 600,
+      fontSize: "24px",
+      lineHeight: "150%",
+      letterSpacing: "-0.5px",
+    },
+    h4: {
+      fontWeight: 500,
+      fontSize: "22px",
+      lineHeight: "150%",
+    },
+    h5: {
+      fontWeight: 500,
+      fontSize: "20px",
+      lineHeight: "150%",
+    },
+    h6: {
+      fontWeight: 500,
+      fontSize: "16px",
+      lineHeight: "120%",
+    },
+    bodyLarge: {
+      fontWeight: 300,
+      fontSize: "18px",
+      lineHeight: "130%",
+    },
+    bodyTitle: {
+      fontWeight: 300,
+      fontSize: "16px",
+      lineHeight: "130%",
+    },
+    bodySmall: {
+      fontWeight: 300,
+      fontSize: "14px",
+      lineHeight: "140%",
+    },
+    bodyTiny: {
+      fontWeight: 300,
+      fontSize: "12px",
+      lineHeight: "120%",
+    },
+  },
+}
+
+export default typography
